Remove mascot prompt elements from Word Bank
