+++ assets/Lessons/exercises/WordBankEnglish/styles.css
@@ -18,85 +18,28 @@
   justify-content: center;
 }
 
-.wordbank__prompt-content {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-}
-
-.wordbank__mascot {
+.wordbank__prompt-lines {
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap: var(--space-sm, 0.75rem);
-  position: relative;
-}
-
-.wordbank__bubble {
+  gap: var(--space-xs, 0.5rem);
+  padding: var(--space-sm, 0.75rem) var(--space-lg, 1.5rem);
+  border-radius: var(--radius-lg, 1.25rem);
   background: rgba(32, 37, 64, 0.96);
   border: 1px solid rgba(108, 125, 197, 0.65);
-  border-radius: 1.25rem;
-  padding: var(--space-sm, 0.75rem) var(--space-lg, 1.5rem);
   box-shadow: 0 10px 28px rgba(8, 12, 34, 0.45);
-  max-width: min(460px, 100%);
-  text-align: center;
-  position: relative;
-  margin-bottom: 2rem;
-  z-index: 1;
+  max-width: min(480px, 100%);
 }
 
-.wordbank__bubble::after {
-  content: '';
-  position: absolute;
-  left: 50%;
-  bottom: -16px;
-  transform: translateX(-50%);
-  border-width: 16px 14px 0 14px;
-  border-style: solid;
-  border-color: rgba(32, 37, 64, 0.96) transparent transparent transparent;
-  filter: drop-shadow(0 6px 12px rgba(8, 12, 34, 0.45));
-}
-
-.wordbank__bubble--no-translit {
-  margin-bottom: var(--space-sm, 0.75rem);
-}
-
-.wordbank__si {
+.wordbank__prompt-si {
   font-size: clamp(1.35rem, 1.1rem + 0.9vw, 1.85rem);
   font-weight: 600;
   letter-spacing: 0.01em;
 }
 
-.wordbank__translit {
-  margin-top: 0.5rem;
+.wordbank__prompt-translit {
   font-size: clamp(0.9rem, 0.85rem + 0.25vw, 1rem);
   color: var(--color-muted, #a9b6c6);
-  text-align: center;
-}
-
-.wordbank__mascot-avatar {
-  width: 96px;
-  height: 96px;
-  border-radius: 1.75rem;
-  background: var(--color-surface, #101a2b);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  box-shadow: var(--shadow-1, 0 8px 24px rgba(0, 0, 0, 0.25));
-  overflow: hidden;
-  font-size: 2.75rem;
-  position: relative;
-  z-index: 2;
-}
-
-.wordbank__mascot-avatar img {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-}
-
-.wordbank__mascot-avatar--fallback {
-  color: #ffd166;
 }
 
 .wordbank__instruction {
@@ -249,7 +192,7 @@
     padding: var(--space-md, 1rem);
   }
 
-  .wordbank__bubble {
+  .wordbank__prompt-lines {
     padding: var(--space-sm, 0.75rem) var(--space-md, 1rem);
   }
